created pipeline demo in notebook
