+++ __init__.py
@@ -1,0 +1,7 @@
+"""HDR Image Grouping Pipeline.
+
+A modular pipeline for grouping HDR images by camera angle,
+with robust handling of exposure variations and real-world photography challenges.
+"""
+
+__version__ = '0.1.0'
