--- experiment_1/model_results/efficientnet_b0_baseline_eps50.json
@@ -1,85 +1,0 @@
-{
-  "metadata": {
-    "model_name": "efficientnet_b0",
-    "eps": 50,
-    "min_samples": 1,
-    "use_two_stage": false,
-    "eps2": null,
-    "use_edge_enhancement": false,
-    "use_perspective_invariant": false,
-    "use_histogram_equalization": true,
-    "use_enhanced_features": true,
-    "runtime_seconds": 43.08695602416992
-  },
-  "clusters": {
-    "1": [
-      "212A2700_1.jpg",
-      "212A2701_1.jpg",
-      "212A2704_1.jpg",
-      "212A2705_1.jpg",
-      "212A2706_2.jpg",
-      "212A2707_2.jpg",
-      "212A2708_2.jpg",
-      "212A2711_2.jpg",
-      "212A2712_3.jpg",
-      "212A2713_3.jpg",
-      "212A2716_3.jpg",
-      "212A2717_3.jpg",
-      "212A2718_4.jpg",
-      "212A2719_4.jpg",
-      "212A2720_4.jpg",
-      "212A2723_4.jpg",
-      "212A2724_5.jpg",
-      "212A2725_5.jpg",
-      "212A2728_5.jpg",
-      "212A2729_5.jpg",
-      "212A2730_6.jpg",
-      "212A2731_6.jpg",
-      "212A2732_6.jpg",
-      "212A2735_6.jpg",
-      "212A2736_7.jpg",
-      "212A2737_7.jpg",
-      "212A2740_7.jpg",
-      "212A2741_7.jpg",
-      "212A2742_8.jpg",
-      "212A2743_8.jpg",
-      "212A2744_8.jpg",
-      "212A2745_8.jpg",
-      "212A2748_8.jpg",
-      "212A2749_9.jpg",
-      "212A2750_9.jpg",
-      "212A2753_9.jpg",
-      "212A2754_9.jpg",
-      "212A2755_9.jpg",
-      "212A2756_9.jpg",
-      "212A2757_10.jpg",
-      "212A2758_10.jpg",
-      "212A2759_10.jpg",
-      "212A2762_10.jpg",
-      "212A2763_11.jpg",
-      "212A2764_11.jpg",
-      "212A2767_11.jpg",
-      "212A2768_11.jpg",
-      "212A2769_11.jpg",
-      "212A2771_11.jpg",
-      "212A2772_11.jpg",
-      "212A2773_11.jpg",
-      "212A2774_11.jpg",
-      "212A2777_11.jpg",
-      "212A2778_12.jpg",
-      "212A2779_12.jpg",
-      "212A2782_12.jpg",
-      "212A2783_12.jpg",
-      "212A2784_12.jpg",
-      "212A2785_14.jpg",
-      "212A2786_14.jpg",
-      "212A2787_14.jpg",
-      "212A2788_14.jpg",
-      "212A2789_14.jpg",
-      "212A2792_14.jpg",
-      "212A2793_15.jpg",
-      "212A2794_15.jpg",
-      "212A2802_15.jpg"
-    ]
-  }
-}
